+++ nmosauth/auth_server/security_service.py
@@ -80,6 +80,7 @@ class SecurityService:
             self.httpServer.started.wait()
 
         if self.httpServer.failed is not None:
+            self._cleanup()
             raise self.httpServer.failed
 
         self.logger.writeInfo("Running on port: {}".format(self.httpServer.port))
@@ -89,7 +90,6 @@ class SecurityService:
         self.start()
         while self.running:
             time.sleep(1)
-        self._cleanup()
 
     def _cleanup(self):
         if self.mdns:
@@ -99,17 +99,16 @@ class SecurityService:
                 self.logger.writeInfo("mDNS stopped gracefully")
             except Exception as e:
                 self.logger.writeWarning("Could not stop mDNS gracefully: {}".format(e))
-
         self.httpServer.stop()
         self.logger.writeInfo("Stopped Http Server")
 
     def sig_handler(self):
-        self.logger.writeInfo('Pressed ctrl+c')
+        print('Pressed ctrl+c')
         self.stop()
 
     def stop(self):
-        self.running = False
         self._cleanup()
+        self.running = False
 
 
 if __name__ == '__main__':